requirements.txt
apache-libcloud>=2.1.0
PyYaml>=3.12
urllib3>=1.22
